added users authorization

+++ app/Http/Controllers/Admin/VideoController.php
@@ -7,7 +7,7 @@ use App\Http\Requests\Admin\VideoCreateRequest;
 use Datatables;
 
 class VideoController extends AdminController
-{    
+{
     /**
      * Create a new controller instance.
      *
@@ -16,8 +16,6 @@ class VideoController extends AdminController
     public function __construct()
     {
         parent::__construct();
-
-        $this->middleware(['auth', 'role:admin']);
     }
     
     /**
@@ -27,51 +25,58 @@ class VideoController extends AdminController
      */
     public function index()
     {
+        $this->authorize('view', Video::class);
         return view('admin.video.list');
     }
 
-    public function list()
+    public function data()
     {
+        $this->authorize('view', Video::class);
         $videos = Video::select(['title', 'published', 'created_at', 'id']);
         return Datatables::of($videos)
-            ->editColumn('published', function ($video) { 
-                    if($video->published) 
-                        return "Published";
-                    else
-                        return "Draft";
-                    })
-            ->editColumn('id', function($video){
-                    return view('admin.video.icons', ['video' => $video])->render();
-                    })
+            ->editColumn('published', function ($video) {
+                if ($video->published) {
+                    return "Published";
+                } else {
+                    return "Draft";
+                }
+            })
+            ->editColumn('id', function ($video) {
+                return view('admin.common._icons', ['model' => $video, 'route' => 'videos'])->render();
+            })
             ->rawColumns(['published', 'id'])
             ->make(true);
     }
     
     public function create()
     {
+        $this->authorize('create', Video::class);
         return view('admin.video.create');
     }
     
     public function show($id)
     {
+        $this->authorize('view', Video::class);
         $video = Video::find($id);
         return view('admin.video.edit', compact('video'));
     }
     
     public function edit($id)
     {
+        $this->authorize('update', Video::class);
         $video = Video::find($id);
         return view('admin.video.edit', compact('video'));
     }
     
     public function store(VideoCreateRequest $request)
     {
+        $this->authorize('create', Video::class);
         try {
             $data = $request->all();
             if (!empty($data['title'])) {
                 unset($data['_token']);
                 Video::create($data);
-                toast('Video created successfully!','success');
+                toast('Video created successfully!', 'success');
             } else {
                 toast('Video Title is must!', 'error');
                 return redirect()->back()->withInput();
@@ -85,6 +90,7 @@ class VideoController extends AdminController
     
     public function update($id, VideoCreateRequest $request)
     {
+        $this->authorize('update', Video::class);
         try {
             $data = $request->all();
             if (!empty($data['title'])) {
@@ -95,7 +101,7 @@ class VideoController extends AdminController
                     $video->$key = $value;
                 }
                 $video->update();
-                toast('Video updated successfully!','success');
+                toast('Video updated successfully!', 'success');
             } else {
                 toast('Video Title is must!', 'error');
                 return redirect()->back()->withInput();
@@ -109,11 +115,12 @@ class VideoController extends AdminController
 
     public function destroy($id)
     {
+        $this->authorize('delete', Video::class);
         try {
             $video = Video::find($id);
             if ($video) {
                 $video->delete();
-            }            
+            }
             return response()->json('Video deleted successfully!');
         } catch (\Exception $ex) {
             return response()->json('Some error occurred!');
